requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
APScheduler==3.10.4
apscheduler-di==0.1.0
asyncpg==0.29.0
attrs==24.2.0
certifi==2024.7.4
chardet==3.0.4
frozenlist==1.4.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
magic-filter==1.0.12
multidict==6.0.5
peewee==3.17.6
psycopg==3.2.1
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.8
rfc3986==1.5.0
rodi==2.0.6
six==1.16.0
sniffio==1.3.1
transliterate==1.10.2
typing_extensions==4.12.2
tzlocal==5.2
yarl==1.9.4
